/lib > /app
Also move main app code to /app folder rather than top-level index.js.
