+++ app/middleware/auth.js
