+++ app/middleware/earlyLogger.js
