+++ app/middleware/logger.js
