+++ app/middleware/static.js
